requirements.txt
amadeus==12.0.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.5
urllib3==2.5.0
